一時保存

+++ GameEngine6/SwapType.h
@@ -10,7 +10,7 @@
 //T_1::タイプ1
 //T_2::タイプ2
 //t_Order::trueなら[T_1],[T_2]を入れ替える
-template<class T_1, class T_2,int t_Order>
+template<class T_1, class T_2,long long t_Order>
 struct IS_Swap_Type
 {
 	template<bool t_Order>
@@ -29,16 +29,16 @@ struct IS_Swap_Type
 	using Type_2 = S_SwapType<!static_cast<bool>(t_Order)>::Type;
 };
 
-template<class T_1, class T_2, int t_Order>
+template<class T_1, class T_2, long long t_Order>
 using IS_Swap_t1 = typename IS_Swap_Type<T_1, T_2, t_Order>::Type_1;
 
-template<class T_1, class T_2, int t_Order>
+template<class T_1, class T_2, long long t_Order>
 using IS_Swap_t2 = typename IS_Swap_Type<T_1, T_2, t_Order>::Type_2;
 
-template<class T, int t_Order >
+template<class T, long long t_Order >
 using IS_Judge = IS_Swap_Type<std::nullopt_t, T, t_Order>;
 
-template<class T, int t_Order>
+template<class T, long long t_Order>
 using IS_Judge_t = typename IS_Judge<T, t_Order>::Type_1;
 
 //仕様
